refactor(testgen): split Behavioral.scala into cohesive sub-modules
Behavioral.scala had grown to 1297 lines mixing four concerns. Split it
along the seams that are actually decoupled (the per-category emitters
share too many helpers to separate; the real boundaries are the larger
self-contained sub-systems):

- GuardSatisfier.scala (315L): the guard seed-dict fixup engine
  (Fix ADT + recognize/topoOrder/collect). Fully self-contained; used
  only via GuardSatisfier.recognize.
- StateMachineTests.scala (519L): transition + status-restriction test
  generation and their shared scaffolding (NonPathInput, request/given
  builders, enum-value lookup). Exposes transitionEmission +
  statusRestriction* as the entry points the orchestrator calls.
- TestFormat.scala (16L): the shared output-formatting helpers
  (invName / prettyOneLine / escapeDocstring), used by both files.
- Behavioral.scala (472L): now just the orchestrator (emitFor /
  testsForOperation), the per-operation category emitters (ensures /
  negative / invariant / temporal) and their build helpers.

Pure code movement + visibility widening (private -> private[testgen]
on the cross-file entry points and their result types). Behaviour
preserved: testgen suite 287/287 green; scalafmt + scalafix clean.
